requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-stdimage==4.0.0
djangorestframework==3.9.1
idna==2.8
Pillow==5.3.0
progressbar2==3.38.0
python-utils==2.3.0
pytz==2018.7
requests==2.21.0
six==1.11.0
urllib3==1.24.1
xlrd==1.2.0
WeasyPrint